Fixed firebase en reworked some code
