+++ src/components/NavBar.js
@@ -2,9 +2,7 @@ import * as React from 'react';
 import {Text, View, Image, StyleSheet} from 'react-native';
 import {NavigationContainer} from '@react-navigation/native';
 import {createBottomTabNavigator} from '@react-navigation/bottom-tabs';
-import CatsScreen from '../screens/CatsScreen';
 import HomeScreen from '../screens/HomeScreen';
-import DogsScreen from '../screens/DogsScreen';
 import ArmorListView from '../screens/ArmorListView';
 import {createNativeStackNavigator} from '@react-navigation/native-stack';
 import ArmorDetailsView from '../screens/ArmorDetailsView';
@@ -59,7 +57,7 @@ function NavBar() {
             <View>
               <Image
                 style={styles.image}
-                source={require('../../src/assets/icons/armor.png')}
+                source={require('../../src/assets/icons/weapon.png')}
               />
             </View>
           ),
@@ -75,8 +73,6 @@ function App() {
         <Stack.Screen name="Home" component={NavBar} />
         <Stack.Screen name="ArmorListView" component={ArmorListView} />
         <Stack.Screen name="ArmorDetailsView" component={ArmorDetailsView} />
-        <Stack.Screen name="Dogs" component={DogsScreen} />
-        <Stack.Screen name="Cats" component={CatsScreen} />
         <Stack.Screen name="ArmorDetails" component={ArmorDetailsScreen} />
         <Stack.Screen name="WeaponListView" component={WeaponListView} />
       </Stack.Navigator>
