--- src/screens/DogsScreen.js
@@ -1,28 +1,0 @@
-import React, {Fragment} from "react";
-import {
-  Button,
-  SafeAreaView,
-  ScrollView,
-  StatusBar,
-  StyleSheet,
-  Text,
-  useColorScheme,
-  View,
-} from 'react-native';
-
-import {
-  Colors,
-  DebugInstructions,
-  Header,
-  LearnMoreLinks,
-  ReloadInstructions,
-} from 'react-native/Libraries/NewAppScreen';
-import DisplayACatImage from '../components/Cat';
-
-export default function DogsScreen({navigation}) {
-  return (
-    <View>
-      <Text>Ecran des chiens </Text>
-    </View>
-  );
-}
