+++ src/screens/HomeScreen.js
@@ -18,72 +18,26 @@ import {
   LearnMoreLinks,
   ReloadInstructions,
 } from 'react-native/Libraries/NewAppScreen';
-import DisplayACatImage from '../components/Cat';
 import {GoogleSignin} from '@react-native-google-signin/google-signin';
-import auth from '@react-native-firebase/auth';
+import auth, {firebase} from '@react-native-firebase/auth';
+import GoogleSignIn from '../components/GoogleSignIn';
+import GoogleSignOut from "../components/GoogleSignOut";
 
-function GoogleSignIn() {
-  GoogleSignin.configure({
-    webClientId:
-      '432126369602-8fbkntp5p15lefulln4f5b5j985bs89t.apps.googleusercontent.com',
-  });
-  async function onGoogleButtonPress() {
-    // Check if your device supports Google Play
-    await GoogleSignin.hasPlayServices({showPlayServicesUpdateDialog: true});
-    // Get the users ID token
-    const {idToken} = await GoogleSignin.signIn();
-
-    // Create a Google credential with the token
-    const googleCredential = auth.GoogleAuthProvider.credential(idToken);
-    // Sign-in the user with the credential
-    return auth().signInWithCredential(googleCredential);
-  }
-  return (
-    <Button
-      title="Google Sign-In"
-      onPress={() =>
-        onGoogleButtonPress().then(() => console.log('Signed in with Google!'))
-      }
-    />
-  );
-}
 export default function HomeScreen({navigation}) {
   return (
     <View>
-      <Test />
+      <GoogleSignIn />
+      <GoogleSignOut/>
     </View>
   );
 }
-function Test() {
-  GoogleSignin.configure({
-    webClientId:
-      '432126369602-8fbkntp5p15lefulln4f5b5j985bs89t.apps.googleusercontent.com',
-  });
-  const SignInWithGoogleAsync = async () => {
-    // Check if your device supports Google Play
-    await GoogleSignin.hasPlayServices({showPlayServicesUpdateDialog: true});
-    // Get the users ID token
-    const {idToken} = await GoogleSignin.signIn();
-
-    // Create a Google credential with the token
-    const googleCredential = auth.GoogleAuthProvider.credential(idToken);
-    // Sign-in the user with the credential
-    const user_sign_in = auth().signInWithCredential(googleCredential);
-    user_sign_in
-      .then(user => {
-        console.log(user);
-      })
-      .catch(error => {
-        console.log(error);
-      });
+function WhosConnected() {
+  const LogConnected = async () => {
+    console.log(auth().currentUser);
   };
-
-  return (
-    <View>
-      <Button title="Sign in With Google" onPress={SignInWithGoogleAsync} />
-    </View>
-  );
+  return <Button title="who's connected  " onPress={LogConnected} />;
 }
+
 
 const styles = StyleSheet.create({
   button_cat: {
